complete 4.21-4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router()
 const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
 const config = require('../utils/config')
 
 blogsRouter.get('/', async (request, response) => {
@@ -20,12 +19,12 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
   const decodedToken = jwt.verify(request.token, config.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -42,8 +41,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+  if (!(request.token && decodedToken.id)) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const id = request.params.id
+  const blogToDelete = await Blog.findById(id)
+
+  if (blogToDelete.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'unauthorized operation' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
